requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.32
passlib==1.7.4
protobuf==3.20.3
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
